Move helper Python inside the tests/integration directory
