+++ tests/integration/artifacts/files/minimal.png
