+++ tests/integration/fixtures/files/clusters/kind-ci.yml
