+++ tests/integration/fixtures/files/clusters/kind.yml
