+++ tests/integration/fixtures/synapse.py
